aider: Apply inline style to HighlightIcon to change color to light-yellow.
# Aider chat conversation:

USER: OK I was able to fix it by adding a height element ot the styles, can you also color the icon a light-yellow instead of grey?

ASSISTANT: To color the `HighlightIcon` a light-yellow instead of grey, we can apply a style directly to the icon. Here's the updated diff:

```diff
--- client/src/components/BookViewer.js
+++ client/src/components/BookViewer.js
@@ ... @@
       <Popover
         id={popoverId}
         open={popoverOpen}
         anchorEl={anchorEl}
         style={{
           position: 'absolute',
           top: `${popoverPosition.top}px`,
           left: `${popoverPosition.left}px`,
           height: '100px' // Assuming this is the line you added for height
         }}
         onClose={() => setAnchorEl(null)}
       >
         <IconButton onClick={handleCreateHighlight}>
-          <HighlightIcon />
+          <HighlightIcon style={{ color: 'lightyellow' }} />
         </IconButton>
       </Popover>
       // ... existing code ...
     </Paper>
   );
 }

 export default BookViewer;
```

This change applies an inline style to the `HighlightIcon` to change its color to light-yellow. Apply this change to see the updated icon color in your application.